Defined a function for part 1
